+++ Sources/FeedKit/Dates/ISO8601DateFormatter.swift
@@ -46,6 +46,7 @@ class ISO8601DateFormatter: DateFormatter {
     }
     
     override func date(from string: String) -> Date? {
+        let string = string.trimmingCharacters(in: .whitespacesAndNewlines)
         for dateFormat in self.dateFormats {
             self.dateFormat = dateFormat
             if let date = super.date(from: string) {